feature: add main screen frame
